requirements.txt
beautifulsoup4
pillow
python-dotenv
requests
